pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipecat-upliftai"
version = "0.1.0"
description = "UpliftAI TTS plugin for Pipecat"
readme = "README.md"
license = "BSD-2-Clause"
requires-python = ">=3.11"
authors = [
    { name = "UpliftAI", email = "engineering@upliftai.org" },
]
keywords = ["pipecat", "tts", "voice", "ai", "upliftai", "urdu"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pipecat-ai>=1.1.0",
    "websockets>=13.1,<16.0",
    "loguru>=0.7",
]

[project.urls]
Homepage = "https://upliftai.org"
Documentation = "https://docs.upliftai.org"
Repository = "https://github.com/uplift-initiative/pipecat-upliftai"
Issues = "https://github.com/uplift-initiative/pipecat-upliftai/issues"

[dependency-groups]
dev = [
    "pytest>=9.0,<10",
    "pytest-asyncio>=1.0,<2",
    "ruff>=0.12,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pipecat_upliftai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
